feat: set up initial Next.js project structure and configuration files.
